requirements.txt
aniso8601==8.0.0
attrs==19.2.0
Click==7.0
Flask==1.1.1
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
protobuf==3.10.0
pyrsistent==0.15.4
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.16.0
